Damage calculation, unithealth, int to float
- Removed the test healthbar scripts, textures and materials.
- Set the Unithealth text to the middle of the unit.
- Changed some int property from building and unit to float values. E.g.
capturepoints, health and loot, player gold.
- The minimum dmg done is 1f the max is float.MaxValue.
- Applying dmg is done via floats. Only the visuals are casted to an
int.
- The range of unit health is still 0 till 10f. Because is does not
matter if we keep it from 0 till 100 or 0 till 10. Since we now hold in
in floats.
- Remade the level so each player has an even amount of units.

+++ Assets/Scripts/Buildings/BuildingGameObject.cs
@@ -31,7 +31,7 @@ public class BuildingGameObject : MonoBehaviour
     public void UpdateCapturePointsText()
     {
         TextMesh text = CapturePointsText.GetComponent<TextMesh>();
-        text.text = BuildingGame.currentCapturePoints + "/" + BuildingGame.capturePoints;
+        text.text = ((int)BuildingGame.currentCapturePoints) + "/" + ((int)BuildingGame.capturePoints);
         CapturePointsText.renderer.enabled = (!Tile.IsFowOn() && BuildingGame.currentCapturePoints > 0);
     }
